+++ TDEngine2/include/ecs/CBaseSystem.h
@@ -33,6 +33,22 @@ namespace TDEngine2
 			TDE2_API void OnInit() override;
 
 			/*!
+				\brief The method adds a command into the deffered executed buffer
+
+				\param[in] action The callback that will be executed later after all systems'll invoke their Update
+
+				\return RC_OK if everything went ok, or some other code, which describes an error
+			*/
+
+			TDE2_API E_RESULT_CODE AddDefferedCommand(const TCommandFunctor& action = nullptr) override;
+
+			/*1
+				\brief The method executes all deffered commands that were accumulated in current frame
+			*/
+
+			TDE2_API void ExecuteDefferedCommands() override;
+
+			/*!
 				\brief The method is invoked when the system is destroyed
 			*/
 
@@ -49,6 +65,7 @@ namespace TDEngine2
 		protected:
 			DECLARE_INTERFACE_PROTECTED_MEMBERS(CBaseSystem)
 		private:
-			bool mIsActive;
+			std::vector<TCommandFunctor> mDefferedCommandsBuffer;
+			bool                         mIsActive;
 	};
 }